Add test case for Opposite property and move test file to Sonar folder
